nvim: add vim-polyglot back
It's not that slow, and some stuff in Python was weird. I could go
hunting for the specific filetype plugins, but I play around with
random languages somewhat frequently.

The problem I run into with some plugins is that they modify expandtab
sometimes, but I can just undo that. Currently doing it with the
editorconfig module, which isn't great, but not too bad.