nvim: use on_init to generate efm's config
I could try to make this async, but using vim.wait actually slows down
a bunch of things, so that's an option. I'd be down to implementing
something that doesn't require vim.wait, but that's going to be callback
hell. Maybe some-day? idk

With my personal tests this seemed fine. In theory the worst part is
parsing .pre-commit-config.yaml and generating the Python config, but
that's pretty fast anyways.