nvim/lsp/formatting: autofmt hack for "bad" servers
Some servers can't detect that the file didn't change and always send a
textEdit, and neovim isn't smart enough to detect that the contents of
the buffer didn't actually change, so here we do something ugly: if the
last time we invoked `update` happened less than
10 milliseconds ago, don't invoke update again, but instead use `noau
update`. I'm not saving a file twice in 10ms :)

This can break with Go where we run source.organizeImports right after
textDocument/formatting, but gopls is a well behaved server, so we
should be fine. I haven't implemented the same logic for
source.organizeImports for that reason.

10ms is arbitrary and works for now, a server that behaves poorly and is
slow could make things worse, and if that happens I'll simply bump up
that 10ms threshold or make it possible to set a threshold per server.