nvim/lsp/efm: fix generation of python tools
Originally only tested this at $WORK, so missed that it didn't quite
work when the pre-commit file wasn't present.

The idea is that we play safe here: if the file doesn't exist, use
default tools, but if the file exists and reading it fails, we simply
don't set any tools.

Also remove the prequire usage with lyaml. It should always be
available.