nvim: adopt nvim-cmp again
With autocompletion enabled. I think I'll probably use this in the long
run, not sure x)

In the past, when I tried it, I hated that everything could get quite
slow with the buffers and also annoying with the popup showing up all
the time. The solution is actually simple: customize keyword_length for
the buffer source! :D