nvim/lsp/references: remove TODO note
This is actually too difficult: language servers return a range, not the
contents of the line, so I'd have inspect the buffers and check the line
contents, which is probably too expensive to do here. Instead what I'm
going to do is continue to suffer, but with fzf making it a bit less
hard.