nvim: revisit the pattern where I shared mod
This was still present in 3 modules: cleanup, packed and shortcut. This
dates back to when the config was in lua, where I followed a pretty
common pattern:

```lua
local M = {state = []}

function M.thing_that_touches_state()
  -- uses M.state
end

return M
```

When I migrated to fennel, I chose a dumb approach for the translation,
and figured I could revisit it later. This is later.