nvim/lsp/buf-diagnostic: add ability to filter diagnostics
The idea is to allow clients to set their own diagnostics filters. This
is probably over kill, but ultimately I want to be able to filter out
diagnostics from pyright that have the "Unnecessary" tag set on them.

A side note on that specifically: that also filters out unused
variables, which isn't really something I'd like to filter out, but the
trick here is that using flake8 I can still catch those! :D