nvim/lsp/buf-diagnostic: fix nil indexing on publishDiagnostics
This was caused because the filter function would return nil whenever
the client was gone, so if there was a race between quitting neovim and
handling publishDiagnostics.

This is likely a side effect of using lib.debounce: the debouncer holds
on to the most recent call, and whenever the timer is stopped it tries
to call the debounced function, which raises the issue.

Closes #973.