nvim/lsp/fs-watch: keep track of registrations instead of client
The order of registrations/unregistrations isn't guaranteed, so we want
to be able uniquely identify the registrations, instead of the
individual clients.

Example from pyright:

1. Receive registration reg-1
2. Receive registration reg-2
3. Receive registration reg-3
4. Receive unregistration reg-1
5. Receive unregistration reg-2

The previous version only kept track of the client-id, meaning that in
step 4 it'd simply disable the file watcher.

One thing to fix in the future is the fact that we don't dedupe events,
but I'll take care of that later once I implement batching.