nvim/lsp/formatting: always write if change-tick changes
Use the hash to determine whether or not we're gonna set noautocmd (love
the double negation).

Ran into this with taplo and pyproject.toml, where it always rewrites
the file even though the hash doesn't change, and then I couldn't exit
neovim without bang because the file had pending writes (but not
really).