nvim/lib/nvim-helpers: wrap the callback again
I want to pass arguments through, but I don't want to return anything.
neovim has made the horrible decision of interpreting the return value
of the callback, and that doesn't make any sense. This led to some of my
autocmds getting unregistered after being executed once, which is really
fun to debug :)