nvim: use a python script to generate efm tools for Python
This is more readable and easier to reason about. Thanks to types, I was
basically able to rewrite this and have it working on the first try.

The Python script isn't perfect as it is currently a direct rewrite
(without callbacks), later I may try to make it more idiomatic.

Closes #1306.