Set required env vars in alacritty.yml
This means we need to generate alacritty.yml, so this commit introduces
a new script to just do that, which isn't too bad. That new script is
also used to activate/deactive demo mode.

Since we're setting those environment variables in alacritty now, we
don't need .zshenv anymore, and we need to be able to keep the current
environment in tmux. That's what the `-L` flag does for tmux.

Finally, now that `.zshenv` isn't a thing anymore, we can just get rid
of the whole link skip logic: the only reason it existed was to prevent
linking stuff to the $HOME dir in a private environment.

The goal is to get rid of as many linking as possible, and right here
we're also getting rid of the linking for tmux.