nvim/langservers: re-gen requirements.txt
This isn't great, but it looks like dependabot only really keeps the
things listed in requirements.in up to date, not the things in
requirements.txt, which is definitely not what I intended.

Perhaps I could have those files in separate dirs, but that's too big of
a hack. Doing this manually every now and then isn't too bad.