nvim: rudimentary integration between jdtls and lombok
I'm not on the work laptop, where I do have a project that uses lombok,
so I'm just gonna get the latest version and always enable it.

I may need to tune this to conditionally include -javaagent and select
the version of lombok based on the project configuration, but let's only
worry about that if it ends up being a real issue.

Closes #1371.