nvim/ftplugin/kotlin: abort the kls: stuff
The problem is that when jumping to Java files, attaching jdtls is
tricky (the reason I'd like to attach jdtls is to keep jumping around).
Here, I had three choices:

1. set `kotlin.externalSources.autoConvertToKotlin` to `true` and
   continue to jump around with the kotlin-language-server
2. hack around to make sure jdtls can attach to the decompiled source
3. disable the feature

I went with (3) because that makes my life easier. Also, since I'm
revisiting the settings, let's configure linting and the JVM target.