nvim: make handling of autofmt and auto-action more uniform
Take the client-id and remove some hacks from the way auto-action was
being handled. In the new approach, if for whatever reason I want to use
separate servers for formatting and organizing imports, I can (for
example, in Java, I could use efm to format code with google-java-format
and still use jdtls to organize imports).