nvim/ftplugin/zig: run `zig build` via efm-langserver
Surprisingly, zls doesn't publish some errors (???), so let's leverage
the lintOnSave option to run `zig build` and capture those errors.

One downside is that it really only lints on save and doesn't lint the
file when it's first opened. Perhaps I can contribute the feature to
efm.

Also, this is totally my fault, but when efm is used, it is preferred
for formatting, so I have to add `zig fmt` too.