tmux: simplify handle-tmux-selection
I originally created it when I wanted to be able to automatically copy
and/or send to nvim, but nowadays I use `y` for copying, which makes
more sense.

Now that `Enter` only means "send to nvim", let's make it official. And
at this point I can write a few lines of shell script (it passes the "no
conditionals" test :D).