nvim/lsp/fs-watch: wrap matching in vim.schedule
This is something that I broke when I migrated from pcre to reduce the
number of deps. The problem is that I used this for a very short amount
of time, and then moved to the builtin.

The builtin one isn't great though, so let's stick with my custom one.